Extend Image/requirements.txt
Flask==2.2.2
gradio-client==0.2.3
Pillow==9.1.0